# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OrbitDesk Enterprise Ticketing System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  super_admin: "super_admin",
  admin: "admin",
  manager: "manager",
  agent: "agent",
  employee: "employee",
  external: "external",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  departmentId?: number | null;
  departmentName?: string | null;
  avatar?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  super_admin: "super_admin",
  admin: "admin",
  manager: "manager",
  agent: "agent",
  employee: "employee",
  external: "external",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
  departmentId?: number | null;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  super_admin: "super_admin",
  admin: "admin",
  manager: "manager",
  agent: "agent",
  employee: "employee",
  external: "external",
} as const;

export interface UpdateUserBody {
  name?: string;
  role?: UpdateUserBodyRole;
  departmentId?: number | null;
  isActive?: boolean;
}

export interface Department {
  id: number;
  name: string;
  description?: string | null;
  color?: string | null;
  icon?: string | null;
  slaResponseHours: number;
  slaResolutionHours: number;
  agentCount: number;
  openTicketCount: number;
  createdAt: string;
}

export interface CreateDepartmentBody {
  name: string;
  description?: string | null;
  color?: string | null;
  icon?: string | null;
  slaResponseHours?: number;
  slaResolutionHours?: number;
}

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  open: "open",
  assigned: "assigned",
  in_progress: "in_progress",
  waiting: "waiting",
  resolved: "resolved",
  closed: "closed",
} as const;

export type TicketPriority =
  (typeof TicketPriority)[keyof typeof TicketPriority];

export const TicketPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface Ticket {
  id: number;
  ticketNumber: string;
  subject: string;
  description: string;
  status: TicketStatus;
  priority: TicketPriority;
  departmentId?: number | null;
  departmentName?: string | null;
  assigneeId?: number | null;
  assigneeName?: string | null;
  createdById: number;
  createdByName: string;
  tags: string[];
  slaBreached: boolean;
  slaDeadline?: string | null;
  commentCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface Comment {
  id: number;
  ticketId: number;
  content: string;
  isInternal: boolean;
  authorId: number;
  authorName: string;
  authorAvatar?: string | null;
  createdAt: string;
}

export interface TicketHistory {
  id: number;
  ticketId: number;
  action: string;
  oldValue?: string | null;
  newValue?: string | null;
  changedById: number;
  changedByName: string;
  createdAt: string;
}

export type TicketDetail = Ticket & {
  comments: Comment[];
  history: TicketHistory[];
};

export type CreateTicketBodyPriority =
  (typeof CreateTicketBodyPriority)[keyof typeof CreateTicketBodyPriority];

export const CreateTicketBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateTicketBody {
  subject: string;
  description: string;
  priority?: CreateTicketBodyPriority;
  departmentId?: number | null;
  assigneeId?: number | null;
  tags?: string[];
}

export type UpdateTicketBodyStatus =
  (typeof UpdateTicketBodyStatus)[keyof typeof UpdateTicketBodyStatus];

export const UpdateTicketBodyStatus = {
  open: "open",
  assigned: "assigned",
  in_progress: "in_progress",
  waiting: "waiting",
  resolved: "resolved",
  closed: "closed",
} as const;

export type UpdateTicketBodyPriority =
  (typeof UpdateTicketBodyPriority)[keyof typeof UpdateTicketBodyPriority];

export const UpdateTicketBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface UpdateTicketBody {
  subject?: string;
  description?: string;
  status?: UpdateTicketBodyStatus;
  priority?: UpdateTicketBodyPriority;
  departmentId?: number | null;
  assigneeId?: number | null;
  tags?: string[];
}

export interface TicketListResponse {
  tickets: Ticket[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateCommentBody {
  content: string;
  isInternal?: boolean;
}

export interface DashboardStats {
  totalTickets: number;
  openTickets: number;
  inProgressTickets: number;
  resolvedTickets: number;
  closedTickets: number;
  urgentTickets: number;
  slaBreachedTickets: number;
  avgResolutionHours: number;
  ticketsCreatedToday: number;
  ticketsResolvedToday: number;
}

export interface DepartmentStat {
  departmentId: number;
  departmentName: string;
  openCount: number;
  resolvedCount: number;
  breachedCount: number;
  color?: string | null;
}

export interface AgentPerformance {
  agentId: number;
  agentName: string;
  avatar?: string | null;
  resolvedCount: number;
  avgResolutionHours: number;
  activeTickets: number;
  satisfactionScore?: number | null;
}

export interface SlaOverview {
  totalWithSla: number;
  compliant: number;
  breached: number;
  atRisk: number;
  complianceRate: number;
}

export type ListTicketsParams = {
  status?: ListTicketsStatus;
  priority?: ListTicketsPriority;
  departmentId?: number;
  assigneeId?: number;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListTicketsStatus =
  (typeof ListTicketsStatus)[keyof typeof ListTicketsStatus];

export const ListTicketsStatus = {
  open: "open",
  assigned: "assigned",
  in_progress: "in_progress",
  waiting: "waiting",
  resolved: "resolved",
  closed: "closed",
} as const;

export type ListTicketsPriority =
  (typeof ListTicketsPriority)[keyof typeof ListTicketsPriority];

export const ListTicketsPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export type ListUsersParams = {
  role?: ListUsersRole;
  departmentId?: number;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  super_admin: "super_admin",
  admin: "admin",
  manager: "manager",
  agent: "agent",
  employee: "employee",
  external: "external",
} as const;

export type GetRecentTicketsParams = {
  limit?: number;
};