# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OrbitDesk Enterprise Ticketing System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{departmentId}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get department by ID
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
...
  /dashboard/recent-tickets:
    get:
      operationId: getRecentTickets
      tags: [dashboard]
      summary: Get recent ticket activity
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, manager, agent, employee, external]
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
...
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        slaResponseHours:
          type: integer
        slaResolutionHours:
          type: integer
        agentCount:
          type: integer
        openTicketCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slaResponseHours
        - slaResolutionHours
        - agentCount
        - openTicketCount
        - createdAt
    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        slaResponseHours:
          type: integer
          default: 4
        slaResolutionHours:
          type: integer
          default: 24
      required:
        - name
    Ticket:
      type: object
      properties:
        id:
          type: integer
        ticketNumber:
          type: string
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, assigned, in_progress, waiting, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        assigneeId:
          type: integer
          nullable: true
        assigneeName:
          type: string
          nullable: true
        createdById:
          type: integer
        createdByName:
          type: string
        tags:
          type: array
          items:
            type: string
        slaBreached:
          type: boolean
        slaDeadline:
          type: string
          format: date-time
          nullable: true
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - ticketNumber
        - subject
        - description
        - status
        - priority
        - createdById
        - createdByName
        - tags
        - slaBreached
        - commentCount
        - createdAt
        - updatedAt
...