# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OrbitDesk Enterprise Ticketing System API
 * OpenAPI spec version: 0.1.0
 */
import type { TicketPriority } from "./ticketPriority";
import type { TicketStatus } from "./ticketStatus";

export interface Ticket {
  id: number;
  ticketNumber: string;
  subject: string;
  description: string;
  status: TicketStatus;
  priority: TicketPriority;
  departmentId?: number | null;
  departmentName?: string | null;
  assigneeId?: number | null;
  assigneeName?: string | null;
  createdById: number;
  createdByName: string;
  tags: string[];
  slaBreached: boolean;
  slaDeadline?: Date | null;
  commentCount: number;
  createdAt: Date;
  updatedAt: Date;
}